refactor: extract credential parser to separate file
